pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arborist"
version = "0.1.0"
description = "Agentic tree search engine for parallelized experiment orchestration"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Jared", email = "jared@arborist.dev" },
]
keywords = ["tree-search", "experiment", "orchestration", "agentic", "mcts"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "litellm>=1.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
arborist = "arborist.cli:cli"

[project.urls]
Homepage = "https://github.com/jared/arborist"
Repository = "https://github.com/jared/arborist"
Issues = "https://github.com/jared/arborist/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
